+++ CHANGELOG.md
@@ -5,7 +5,7 @@
   * **Enhancement** Added `RefreshTokenType` to the `ApplicationOIDC` data model. [#518](https://github.com/patrickcping/pingone-go-sdk-v2/pull/518)
   * **Enhancement** Added `PasscodeGracePeriod` and `Push` to the `ApplicationOIDCAllOfMobile` data model. [#518](https://github.com/patrickcping/pingone-go-sdk-v2/pull/518)
   * **Enhancement** Added `EnableAlwaysAcceptAcsUrlInSignedAuthnRequest` to the `ApplicationSAML` data model. [#518](https://github.com/patrickcping/pingone-go-sdk-v2/pull/518)
-  * **Enhancement** Added `WsTrustVersion` to the `ApplicationWSFED` data model. [#518](https://github.com/patrickcping/pingone-go-sdk-v2/pull/518)
+  * **Enhancement** Added `AssertionDuration` and `WsTrustVersion` to the `ApplicationWSFED` data model. [#518](https://github.com/patrickcping/pingone-go-sdk-v2/pull/518)
   * **Enhancement** Added `ExternalId` to the `ApplicationExternalLink` data model. [#518](https://github.com/patrickcping/pingone-go-sdk-v2/pull/518)
   * **Enhancement** Added `TOKEN_EXCHANGE` and `CIBA` values to the `EnumApplicationOIDCGrantType` enum type. [#518](https://github.com/patrickcping/pingone-go-sdk-v2/pull/518)
   * **Enhancement** Added `PayloadOptions` to the `Subscription` data model. [#519](https://github.com/patrickcping/pingone-go-sdk-v2/pull/519)